refactor: seperate calendar logic
