+++ src/components/recordDashboard/Calendar.tsx
@@ -6,6 +6,7 @@ import {
 } from "@tabler/icons-react";
 import React, { useState } from "react";
 import Image from "next/image";
+import calendarLogic from "@/common/helpers/calendarLogic";
 
 interface DataType {
   card: string;
@@ -281,26 +282,6 @@ const Calendar = () => {
       "Fri",
       "Sat",
     ];
-    const calendarLogic = () => {
-      const calendarArray = [];
-      const totalDays = 42;
-
-      // 取得當前月份第一天的日期
-      const firstDayOfMonth = moment().startOf("month");
-      // 取得當前月份第一天所在的那一週的第一天(星期天)的日期
-      const firstDateOfCalendar = firstDayOfMonth.clone().startOf("week");
-
-      let eachWeek = [];
-      for (let i = 0; i < totalDays; i++) {
-        const currentDate = firstDateOfCalendar.clone().add(i, "days");
-        eachWeek.push(currentDate.format("YYYY-MM-DD"));
-        if (eachWeek.length === 7) {
-          calendarArray.push(eachWeek);
-          eachWeek = [];
-        }
-      }
-      return calendarArray;
-    };
 
     const calendarArray: string[][] = calendarLogic();
     const allDates: React.ReactNode[] = [];
